encode conversion
This file shows some practice how  some special character are deal with in server side of web system including no-width space, control character in order to be recognized by different encoding. Especially, there are some old characters in Japanese Windows are not able to be recognized by Latin-based encoding, so how to convert them or display them in HTML  is one of the purposes of writing this file.
Some HTML based library are used and some sample code are comment out.