People saying 3 or times of 3 out of circle
Issue dicription: there are 50 people who call out numbers from 1 in loop 
(after the 50th people said 50, the first peple say 51). The people will be out of the loop 
if his number is 3 or times of 3. The output will be the last people.
example input : 50  output: 11

next version: modify the code to accept input from console and output the result